Generate field getters in macro
I've looked into available solutions for this kind of thing, but I
couldn't find anything that fits our needs. The bitfield crate comes
closest, but it doesn't support byte slices as a type for the bitfield.
I've come to the conclusion that creating something that's purpose-built
for our needs is the best solution for now.

I haven't put much thought into the performance of this, but I assume
that a lot of stuff will be optimized out in release mode, since it
depends on constant values.