Rename `ci` directory to `scripts`
The directory is not only for CI, but can also be used locally, so the
old name was too specific. In addition, I have other projects where it's
called `scripts`, and I don't want to complicate things for myself.

A note about the script itself: It might seem nonsensical to have such a
simple script, but only if you consider this project in isolation. Fact
is, I have multiple projects that I need to build regularly, and without
a script, I'd have to remember which project has examples that I can't
forget to build, etc.