Provide full access to TX_BUFFER data
The previous approach was not flexible enough for any but the most basic
use cases. The new approach is just to return a mutable reference to the
buffer, which is the most flexible approach.

Unfortunately this new approach doesn't play well with the `read`/
`write`/`modify` closures, which were designed with the fluent API of
the other registers in mind, but I think this is the right trade-off in
this case.