Generate `ToBytes` and `FromBytes` impls in macro
This reduces redundancy, but as a side effect, it also makes `FromBytes`
more flexible. It's no longer required to pass a slice that is exactly
as long as the value that's being produced. If a slice is shorter, the
missing bytes are being treated as `0`.

This side effect is quite important, actually, because there are
registers that are shorter than some of the types used for representing
their fields. Without this change, those registers can't be supported.