Add module to support double-sided two-way ranging
I have example programs running on top of this module, and while they
work as intended overall, this doesn't currently result in an accurate
range measurements. This can be attributed to problems with these
example programs that I have yet to fix, but there is the possibility
this module also contains bugs that contribute to the inaccurate range
measurement.