Depend on Git version of nrf52832-hal
My next step towards replacing the depency on nrf52832-hal with
embedded-hal is to port over SPI. The latest release of nrf52832-hal
doesn't implement the embedded-hal traits for SPI, so the latest Git
version is required to continue here.

I don't think this is going to be a problem, as I plan to remove the
dependency on nrf52832-hal completely, before submitting a pull request.