Remove automatic conversion to `nb::Error`
It's fine as it is right now, but it's causing problems for a change I'm
working on. `Error` will gain a type parameter (required for
embedded-hal use). With that type parameter, that `From` implementation
no longer compiles, and I don't think it's worth the trouble to figure
out whether that can be fixed somehow.