Replace `TxFuture` with new `Sending` state
Instead of `send` returning a `TxFuture` which borrows `DW1000`, this
moves `DW1000` into `send`, which then returns an updated `DW1000`
instance, with the new `Sending` type state.

The reason for this change is, that the borrowing is too restrictive and
will prevent `DW1000` from being used as a struct field. In that case,
the `TxFuture` would also have to be stored in the struct, which is not
possible due to self-borrowing.

This is only the first commit in a series. I'm going to replace
`RxFuture` in the same way, as well as do some cleanup.