Fix linker errors when building examples
References to the interrupt handlers weren't found, because two
different `cortex-m-rt` versions were used. The `device.x` from the PAC
was probably no used, causing the errors.

This wasn't caught by CI (or running the build script locally), because
it worked, if `RUSTFLAGS` was set, even to an empty value. I don't know
why that is.